arrange io image animation into assets

+++ iTag/Door/NFCDoorAccessViewController.m
@@ -700,7 +700,7 @@ UIView *alertViewDoor;
         [self checkKeyStorage];
         [self setNFCData];
     }
-    NSLog(@"~~key :\ncount : %ld\n1 : %@\n2 : %@\n3 : %@\n4 : %@", NFCArray.count, [NFCArray objectAtIndex:0], [NFCArray objectAtIndex:1], [NFCArray objectAtIndex:2], [NFCArray objectAtIndex:3]);
+    NSLog(@"~~key :\ncount : %ld\n1 : %@\n2 : %@\n3 : %@\n4 : %@", (unsigned long)NFCArray.count, [NFCArray objectAtIndex:0], [NFCArray objectAtIndex:1], [NFCArray objectAtIndex:2], [NFCArray objectAtIndex:3]);
 }
 
 #pragma mark door BTSmartSensorDelegate
